psql tuples.
Former-commit-id: ce055b4461850ded8b86e986bee286e9b90eb0bf

+++ src/backend/catalog/schema.h
@@ -290,11 +290,10 @@ class Schema	{
   void AddConstraintByColumnId(  oid_t column_id, catalog::Constraint* constraint ) {
     columns[column_id].AddConstraint(constraint);
   }
+
   void AddConstraintByColumnName(  std::string column_name, catalog::Constraint* constraint ) {
-    for( int column_itr = 0; column_itr < columns.size(); column_itr++)
-    {
-      if( columns[column_itr].name == column_name )
-      {
+    for(size_t column_itr = 0; column_itr < columns.size(); column_itr++){
+      if( columns[column_itr].name == column_name ) {
         columns[column_itr].AddConstraint(constraint);
       }
     }
